Automatically pick up new config changes
We need to run write_config.py regularly to pull in new extensions/skins
from Gerrit and GitHub. For now we'll do this as a crontab under my
shell account.

Then we also need to restart hound to pick up those config changes. We can
use systemd to do this, since it already has functionality for timeouts
and automatic restarting.

Change-Id: I19d3a1cfb2e8921682c127e681142ccf87becc5d