Index operations/puppet
puppet is special because its default branch is `production`, not `master`.
And hound hardcodes the latter (https://github.com/etsy/hound/issues/9).

So we set up our own manual clone of the repo (/srv/puppet), and use the
awesome git-symbolic-ref(1) to alias production as master in all the places
that it matters for hound.

That directory is mounted into all the docker containers as /puppet, and hound
is instructed to index and pull that instead of setting up its own clone.

Bug: T211999
Change-Id: Ib0a16521d16745492f6acc530bd6ce5e97e9f065