Introduce wait.py to avoid bootstrapping overloads
We have a somewhat unique startup problem, in that we don't want all of
the hound instances starting up at once because the initial startup is
extremely memory and CPU intensive.

We can't just run one systemd service after another because hound has
started up, so the systemd service is considered ready even though
startup is still going on.

So, introduce wait.py which hits a new JSON health endpoint to see if
anything else is starting up, and if so, waits. We introduce some
randomness so hopefully we don't see a race if the service starts up.

Internally we add a new /_health.json endpoint and introduce a new
'pre-start' status for when the service is waiting with wait.py.

Change-Id: Iea6023664b84d0f20c0c1a1d08def08cb38cc4ac