Add new beta frontend
To replace hound's not very flexible/adaptable one from upstream.

So far this adds features such as:
* Not losing state when switching between search profiles
* Generating a Phabricator checklist based on results
* A table of contents sidebar for a quick overview

Not all features were ported over, especially those that I don't use
personally. If people want them, I can add them back in.

For now I'll set this up on a separate subdomain so people can test and
provide feedback until we feel it's stable enough to replace the current
frontend.

Also I'm not sure the current Rust/WASM build is going to be fast enough
from a performance perspective. Might want to do most of the
work/rendering on the server, which would be pretty simple.

Change-Id: I0455e8b023d7fce144654d972418024c2ed0cd82