Have write_config.py restart hound services as needed
A new --restart flag will enable write_config.py to tell systemd to
restart hound-{name} services if the URLs of Git repositories have
changed.

Ideally this will cut down on the number of time hound instances need to
be restarted.

This will require the script be run as root.

Bug: T268198
Change-Id: Ic6c31e789f911ee5d28e9de5183fb476441cb4d8