frontend: un-stick repolist on narrow viewports
Previously, after the first page of scrolling, it peeked through
behind the search results because it was positioned sticky to the
top of the viewport even on mobile. This is fine on wider viewports
where the element is (through another classname) made into a second
column on the side, but making it sticky when it isn't on the side,
means it starts overlapping between results.

Keep the limited height and scrollbar though so that on mobile you
only have to scroll past one page of this before the text results
appear.

Change-Id: I2afaaddac6344b7ad05a8375c5dd1360dd544ec3