write_config: Create 'devtools' backend (split from 'operations')
== Increase index ==

Widen the index from a few handpicked integration/ repos to all
non-archived repos in Gerrit under this prefix, such as:

- integration/commit-message-validator
- integration/docroot
- integration/utils
- integration/visualdiff

== Decrease/maintain clutter level ==

Also avoid increasing UI clutter by hiding a few backends from the
default form (they remain reachable by permalink) that are subsets
of other presets that seem fine to use instead:

* extensions -> extensions & skins (subset)
  skins -> extensions & skins (subset)

  Effectively the same kind of repo and dependency relation.
  Querying extensions without skins or vice versa seems likely to
  either miss important hooks or callers, or to have no false
  positives anyway (when searching for something only skins or only
  extensions do), and they're both pretty fast.

  Especially now with the new typeahead repo filter, the length
  of the repo list itself doesn't play much of a role anymore
  within the form.

* OOUI -> libraries (subset)
  Milkskake -> libraries (subset)

* services -> MW deployed (subset)

Remove some labels that needlessly increased the number of words
for the same thing and decreased familiarity and consistent exposure.

Skins and extensions under the "deployed" preset already used repo
name as labels. Apply this to the "Extensions & skins" preset as well
for consistency.

Keep codesearch-old unchanged, with the new preset hidden there
to avoid line wrapping, and the old settings for which are visible
and which not.

Bug: T303434
Change-Id: I84900141f18f24cf5c517d740b3dd9841055a32e