switch to main gerrit server instead of using the replica
To avoid downtime for codesearch during upcoming maintenance
downtime of gerrit-replica.wikimedia.org suggesting to switch
it to use the main gerrit.wikimedia.org instead, unless there
are concerns with that.

Patch created with:

 sed -i 's/gerrit-replica/gerrit/g' test_write_config.py write_config.py

The current backend of gerrit.wikimedia.org, gerrit1003, is also
new hardware and we just switched to that last week as the production
server, replacing a 4 year old server, gerrit1001.

The reason that we put some clients directly on gerrit-replica
back in the days was to reduce load on the main server but as far
as I can tell this has been fixed in a previous Gerrit version
even regardless of the new hardware and hasn't been an issue for a while.

But the price we pay for it is that it creates real downtime just
because a once-passive host goes down.

For this and other reasons mentioned on wikitech-l today, I also
suggest to just leave it like this forever.

But this can be seen as a slightly separate discussion.

Let me know what you think of it both with or without a revert after the
maintenance on the replica is over.

Bug: T336710
Change-Id: I6f4b8a4e91966b7427264c96f7bbef54ecbc6cb9