frontend: Fix excludeFiles mismatch between form read and URL write
The form submission and reading logic, as well as Hound, and the
backend API, all use `excludeFiles`.

In the frontend, when generating the clickable buttons that preserve
the form submission when navigating to a different search preset,
were setting excludeFilePath=$excludeFiles, thus very much preserving
it, but not in a way that anything reads.

1. https://codesearch.wmcloud.org/core/
2. input "Krinkle", exclude "js", submit
3. https://codesearch.wmcloud.org/core/?q=Krinkle&files=&excludeFiles=js
4. click "Puppet"
5. https://codesearch.wmcloud.org/puppet/?q=Krinkle&excludeFilePath=js

Preserved, and yet.. useless.

Set the correct parameter instead.

Bug: T321347
Change-Id: I179ec1e3d7569f704f6c71caf91e36a7d4fdde1c