frontend: change Issue tracker link from "Create form" to "workboard"
The workboard offers a "New Task" link in the sidebar to create new
tasks. On the other hand, the task creation form is 1) a dead-end
when not logged in e.g. mobile browsing, and 2) doesn't link back
to the workboard even when logged-in, because the "tags" in the
pre-filled form are not hyperlinked clickable.

Change-Id: Icb9a2858ec4e889dec9555d1bd4e6f4ad640a5ba