Enable CORS for app.py errors, and display in frontend
Today, when the Hound instance is starting up, if the frontend has
a php-apcu cache miss, all is good, because the PHP frontend (not subject
to CORS) will notice the Hound instance is not yet up, and render
a helpful message.

But, if the PHP frontend recently saw Hound being up, but it is no
longer up (i.e. every day around UTC mightnight for ~15 min), and
the JS is trying to query it via the codesearch-backend app proxy,
the JS fetch fails with a NetworkError due to CORS not being enabled
on the error pages.

I considered adding `headers=[('access-control-allow-origin', '*')]`
in various places, but then I realized this still would not handle
the case of a 404 Error and a few other likely edge cases, so went
with global middleware instead.

Change-Id: I6126934c67788e43d6c5351c145b086e878e3960