frontend: Add optional CODESEARCH_HOUND_BASE for local Hound API
During the ocasional cache miss for the repo metadata (especially
the "Excludes files" page), there is chance of tripping the CDN
rate limit when multiple users warm up 2+ backends at the same time,
because all requests are going through the WMCS dynamicproxy at
codesearch-backend.wmcloud.org, and are originating from the
cloudsearch8 VM, regardless of which real user it is for.

The custom throttling in Codesearch.php to prevent rate limiting works
fine (and is required) when running it locally as a single user. But,
in production this doesn't work since it only prevents a single web
request from exceeding 100 req/s, it doesn't prevent the server as a
whole from tripping the rate limit.

Instead of complicating this further, do the same as what the old
app.py frontend did, which is talk to the Hound instance direclty,
it literaly runs on the same codesearch8 server on a different port,
we don't need CDN rate limiting there.

Note that this doesn't not expose unthrottled access. The end-user
request to codesearch-frontend still comes in with the same CDN
ratelimit. And the PHP frontend does not allow users to trigger
arbitrary requests. Each frontend query maps to 1 backend query,
with the exception of repo metadata which instead fans out to a short
and finite hardcoded list URLs to fetch metadata from, and is then
cached in APCU (unconditionally, 1 hour). No user input can vary
or increase the N requests URLs.

See also If89be4a1702623af441fdcea0c5969a951252fe8 in Puppet, which
can later activate this logic.

Change-Id: I5ab347f3811cc97e03ca7748bab6788faf339334