app.py: remove DESCRIPTIONS from old UI, add "Hound" to title
Remove the need for app.py to maintain a complete list of backends
and (their descriptions) and iterate `app.config['PORTS']` instead.

This removes the need to keep more things in sync, given that
frontend/ has to do this already.

Also mark this UI more explicitly as being "Hound", treating it more
like a thin wrapper to access the Hound API, with (for now) the
built-in UI, but not with additional features now-replaced by
codesearch-frontend.

Follows-up c913c63976 which removed overrides for favicon and
open_search.

Change-Id: Ifb2063e50c39b404a730376d58e8833f42a78754