frontend: Fix files count in query results stats; rework structure
The performance numbers of a query result include a field named "files"
that one might assume to be the number of returned files,
but actually came from the `Stats.FilesOpened` field of the Hound API
which is an internal counter independent of the number of files found.

That said, the API's output does include the number of files found;
it just happens to be split per repository,
in the `Results.*.FilesWithMatch` fields.

This change implements logic to sum those fields into a single total
and shows that in the "files" field of the query statistics.

The change also reworks the structure and labels of the performance data
to make it more readable and intuitive.

Bug: T364266
Change-Id: I8168ceaf7ceffb91d6020bbb1aa3c8c52004129a